backend/requirements.txt
langgraph>=0.2.40
langchain-groq>=0.2.0
python-dotenv>=1.0.1
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
